+++ packages/code-morph/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @baseplate-dev/code-morph
 
+## 0.6.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @baseplate-dev/utils@0.6.3
+
 ## 0.6.2
 
 ### Patch Changes