+++ packages/core-generators/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @baseplate-dev/core-generators
 
+## 0.6.3
+
+### Patch Changes
+
+- Updated dependencies [[`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931)]:
+  - @baseplate-dev/sync@0.6.3
+  - @baseplate-dev/utils@0.6.3
+
 ## 0.6.2
 
 ### Patch Changes