+++ plugins/plugin-storage/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @baseplate-dev/plugin-storage
 
+## 0.6.3
+
+### Patch Changes
+
+- [#835](https://github.com/halfdomelabs/baseplate/pull/835) [`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931) Thanks [@kingston](https://github.com/kingston)! - Switch generated UI components from radix-ui to @base-ui/react
+  - Replace radix-ui and cmdk dependencies with @base-ui/react in generated projects
+  - Update all generator templates to use base-ui component implementations
+  - Convert asChild prop usage to render prop pattern in admin layout, crud list, and crud edit action generators
+  - Update plugin-storage upload component templates to use Field/FieldError instead of FormItem/FormMessage
+
+- Updated dependencies [[`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931), [`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931), [`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931), [`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931), [`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931), [`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931), [`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931), [`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931), [`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931), [`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931), [`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931), [`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931), [`8072019`](https://github.com/halfdomelabs/baseplate/commit/8072019942287ee35720520d08f517272510c931)]:
+  - @baseplate-dev/react-generators@0.6.3
+  - @baseplate-dev/ui-components@0.6.3
+  - @baseplate-dev/fastify-generators@0.6.3
+  - @baseplate-dev/project-builder-lib@0.6.3
+  - @baseplate-dev/plugin-queue@0.6.3
+  - @baseplate-dev/sync@0.6.3
+  - @baseplate-dev/core-generators@0.6.3
+  - @baseplate-dev/utils@0.6.3
+
 ## 0.6.2
 
 ### Patch Changes